fixed pointer wraparound

+++ vector.h
@@ -1,10 +1,7 @@
 #pragma once
 
-#include <stdlib.h>
-#include <string.h>
 #include <stdbool.h>
-#include <assert.h>
-#include <stdio.h>
+#include <stdint.h>
 
 #ifndef VECTOR_DEFAULT_SIZE
 #define VECTOR_DEFAULT_SIZE 10
@@ -17,33 +14,33 @@
 typedef struct vector
 {
     void* data;
-    unsigned int element_size;
-    unsigned int size;
-    unsigned int capacity;
+    uint64_t element_size;
+    uint64_t size;
+    uint64_t capacity;
 } vector;
 
-vector vector_create_impl(unsigned int element_size);
+vector vector_create_impl(uint64_t element_size);
 
-void vector_push_back_amount(vector* vector, void* item, unsigned int amount);
+void vector_push_back_amount(vector* vector, void* item, uint64_t amount);
 
 // New memory will be zeroed
-void vector_resize(vector* vector, unsigned int size);
+void vector_resize(vector* vector, uint64_t size);
 
-void vector_reserve(vector* vector, unsigned int capacity);
+void vector_reserve(vector* vector, uint64_t capacity);
 
-void vector_change_capacity(vector* vector, unsigned int capacity);
+void vector_change_capacity(vector* vector, uint64_t capacity);
 
 void vector_pop_back(vector* vector);
 
-void vector_erase_index(vector* vector, unsigned int index);
+void vector_erase_index(vector* vector, uint64_t index);
 
 void vector_clear(vector* vector);
 
 void vector_destroy(vector* vector);
 
-void* vector_get(vector* vector, unsigned int index);
+void* vector_get(vector* vector, uint64_t index);
 
-void vector_set(vector* vector, unsigned int index, void* item);
+void vector_set(vector* vector, uint64_t index, void* item);
 
 typedef struct vector_iter
 {
@@ -66,6 +63,6 @@ void vector_iter_dec(vector_iter* iter);
 void* vector_iter_get(vector_iter iter);
 
 // Serialisation
-char* vector_serialise(const vector* vector, unsigned int* data_size);
+char* vector_serialise(const vector* vector, uint64_t* data_size);
 
-vector vector_deserialise(const char* data, unsigned int data_size);
+vector vector_deserialise(const char* data, uint64_t data_size);
